SUP-70 alteracoes estruturais no dashboard + sidebar
